+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+## 0.1.6
+- Added property `base` to class `qarray`.
+
 ## 0.1.5
 - Removed the class `Quantity` and replaced it by a
   function with name `quantity`.
 - Units and unit expressions can now be combined with qarrays
-to create new expressions. 
+to create new expressions.
 
 ## 0.1.4